+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 ## 21.11.2025 `v1.9.1`
 * Unified the module and class docstring styles throughout the whole library.
 * Moved the Protocol `ProgressUpdater` from the `console` module to the `types` module.
+* Added throttling to the `ProgressBar` update methods to impact the actual process' performance as little as possible.
 
 
 <span id="v1-9-0" />